+++ integration/containerd_image_test.go
@@ -20,15 +20,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	goruntime "runtime"
-	"strings"
 	"testing"
 	"time"
 
 	containerd "github.com/containerd/containerd/v2/client"
 	"github.com/containerd/containerd/v2/integration/images"
 	"github.com/containerd/containerd/v2/internal/cri/labels"
-	"github.com/containerd/containerd/v2/pkg/deprecation"
 	"github.com/containerd/containerd/v2/pkg/namespaces"
 	"github.com/containerd/errdefs"
 	"github.com/stretchr/testify/assert"
@@ -265,43 +262,4 @@ func TestContainerdSandboxImagePulledOutsideCRI(t *testing.T) {
 
 	t.Log("ensure correct labels are set on pause image")
 	assert.Equal(t, "pinned", pauseImg.Labels()["io.cri-containerd.pinned"])
-}
-
-func TestContainerdImageWithDockerSchema1(t *testing.T) {
-	t.Setenv(deprecation.EnvPullSchema1Image, "1")
-	if goruntime.GOOS == "windows" {
-		t.Skip("Skipped on Windows because the test image is not a multi-platform one.")
-	}
-
-	var testImage = images.Get(images.DockerSchema1)
-	digest := strings.Split(testImage, "@")[1]
-	ctx := context.Background()
-
-	t.Logf("make sure the test image doesn't exist in the cri plugin")
-	i, err := imageService.ImageStatus(&runtime.ImageSpec{Image: testImage})
-	require.NoError(t, err)
-	if i != nil {
-		require.NoError(t, imageService.RemoveImage(&runtime.ImageSpec{Image: testImage}))
-	}
-
-	t.Logf("pull the image into containerd")
-	//nolint:staticcheck // Ignore SA1019. Need to keep deprecated package for compatibility.
-	_, err = containerdClient.Pull(ctx, testImage, containerd.WithPullUnpack, containerd.WithSchema1Conversion)
-	require.NoError(t, err)
-	defer func() {
-		// Make sure the image is cleaned up in any case.
-		if err := containerdClient.ImageService().Delete(ctx, testImage); err != nil {
-			assert.True(t, errdefs.IsNotFound(err), err)
-		}
-		assert.NoError(t, imageService.RemoveImage(&runtime.ImageSpec{Image: testImage}))
-	}()
-
-	imgByRef, err := containerdClient.GetImage(ctx, testImage)
-	require.NoError(t, err)
-
-	t.Logf("the image should be marked as managed")
-	assert.Equal(t, "managed", imgByRef.Labels()["io.cri-containerd.image"])
-
-	t.Logf("the image should be marked as dokcker schema1 with its original digest")
-	assert.Equal(t, digest, imgByRef.Labels()["io.containerd.image/converted-docker-schema1"])
 }