+++ internal/cri/util/references.go
@@ -40,15 +40,14 @@ func ParseImageReferences(refs []string) ([]string, []string) {
 }
 
 // GetRepoDigestAndTag returns image repoDigest and repoTag of the named image reference.
-func GetRepoDigestAndTag(namedRef reference.Named, digest imagedigest.Digest, schema1 bool) (string, string) {
+func GetRepoDigestAndTag(namedRef reference.Named, digest imagedigest.Digest) (string, string) {
 	var repoTag, repoDigest string
 	if _, ok := namedRef.(reference.NamedTagged); ok {
 		repoTag = namedRef.String()
 	}
 	if _, ok := namedRef.(reference.Canonical); ok {
 		repoDigest = namedRef.String()
-	} else if !schema1 {
-		// digest is not actual repo digest for schema1 image.
+	} else {
 		repoDigest = namedRef.Name() + "@" + digest.String()
 	}
 	return repoDigest, repoTag