+++ pkg/deprecation/deprecation.go
@@ -21,8 +21,6 @@ type Warning string
 const (
 	// Prefix is a standard prefix for all Warnings, used for filtering plugin Exports
 	Prefix = "io.containerd.deprecation/"
-	// GoPluginLibrary is a warning for the use of dynamic library Go plugins
-	GoPluginLibrary Warning = Prefix + "go-plugin-library"
 	// CRIRegistryMirrors is a warning for the use of the `mirrors` property
 	CRIRegistryMirrors Warning = Prefix + "cri-registry-mirrors"
 	// CRIRegistryAuths is a warning for the use of the `auths` property
@@ -42,7 +40,6 @@ const (
 )
 
 var messages = map[Warning]string{
-	GoPluginLibrary: "Dynamically-linked Go plugins as containerd runtimes are deprecated since containerd v2.0 and removed in containerd v2.1.",
 	CRIRegistryMirrors: "The `mirrors` property of `[plugins.\"io.containerd.grpc.v1.cri\".registry]` is deprecated since containerd v1.5 and will be removed in containerd v2.1." +
 		"Use `config_path` instead.",
 	CRIRegistryAuths: "The `auths` property of `[plugins.\"io.containerd.grpc.v1.cri\".registry]` is deprecated since containerd v1.3 and will be removed in containerd v2.1." +