+++ RELEASES.md
@@ -130,7 +130,7 @@ to all committers.
 | [1.6](https://github.com/containerd/containerd/releases/tag/v1.6.0)  | LTS           | February 15, 2022              | July 23, 2025                  | @containerd/committers |
 | [1.7](https://github.com/containerd/containerd/releases/tag/v1.7.0)  | LTS           | March 10, 2023                 | March 10, 2026                 | @containerd/committers |
 | [2.0](https://github.com/containerd/containerd/releases/tag/v2.0.0)  | Active        | November 5, 2024               | November 7, 2025 (_tentative_) | @containerd/committers |
-| [2.1](https://github.com/containerd/containerd/milestone/48)         | Beta          | May 7, 2025 (_tentative_)      | _TBD_                          | _TBD_                  |
+| [2.1](https://github.com/containerd/containerd/milestone/48)         | RC            | May 7, 2025 (_tentative_)      | _TBD_                          | @samuelkarp, @dmcgowan |
 | [2.2](https://github.com/containerd/containerd/milestone/49)         | _Future_      | November 5, 2025 (_tentative_) | _TBD_                          | _TBD_                  |
 
 ### Kubernetes Support